MIPS: JZ4740: A320: Updated signature of panic blink function
For some unspecified reason the time argument changed from a long to an int in 2.6.36.

+++ arch/mips/jz4740/board-a320.c
@@ -51,7 +51,7 @@
  * TODO(MtH): This should use the backlight driver instead of directly
  *            manipulating the GPIO pin.
  */
-static long a320_panic_blink_callback(long time)
+static long a320_panic_blink_callback(int time)
 {
 	gpio_direction_output(JZ_GPIO_PORTD(31), (time / 500) & 1);
 	return 0;
